Day cycle, level transports
Sun cycle added. Goals at the end of levels transport to the next area.

I think that the player should start in the hub at the beginning of the game. They get the first quest for three collectables. They get the climb ability at the end of the tutorial. Return to the hub, turn in the quest, get a new quest, go to climbing level ....ect. the hub is bypassed for now until quests are working. I'll share this template on Trello after sleep. haha.