Area_1_Map1 - Tile Map Absent from Scene Troubleshoot.
The texture, tile set, and tile map were appropriately assigned, I believe. That did not fix the issue of the invisible tile map. But I think that's it's worth saving.

The player has an invisible platform that they are standing on at the beginning of play. Not sure why if the tile map is not there. Move player left or right, they fall off platform.