Key to unlock door!!!
We have a working key to "unlock" the door. The key can be walked over to pick. This might change to an input after I get some sleep. If you pick up the key and press the interact key next to the door, the door will be destroyed. If you press the interact button next to the door after jumping over the key to bypass collecting it, There will be a prompt that tells you to come back later.