Things and Stuff
Added right shift input, controls pages updated accordingly.
No more bouncing off of collectables. I am in the middle of making a dialog system. I am doing this to solve a few issues. A dialog system will allow the player to go outside using the soldiers as a gate. " yes I want to leave" or " no I don't want to leave" The player will be able to tell when they are going to leave. This will cushion the transition from level to level too which was a critique.  I think a dialog system outside of quests is important to progress.